requirements.txt
# --- Core Flask stack ---
Flask>=2.0.0
Flask-Login>=0.5.0
Flask-Mail>=0.9.0
Flask-Migrate>=3.0.0
Flask-SQLAlchemy>=2.5.0
Flask-WTF>=1.0.0
Jinja2>=3.0.0
Werkzeug>=2.0.0
click>=8.0.0
itsdangerous>=2.0.0
python-dotenv>=0.19.0

# --- Database ---
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0
alembic>=1.7.0

# --- Automation ---
playwright>=1.20.0

# --- Machine Learning & Data ---
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.0.0

# --- Excel support ---
openpyxl>=3.0.0

# --- Scheduling ---
APScheduler>=3.7.0

# --- Authentication ---
PyJWT>=2.0.0
